Include customElement.define cost in LitElement and Polymer benchmarks.
