+++ lit-element/shack/index-1.js
@@ -9,13 +9,5 @@
  * rights grant found at http://polymer.github.io/PATENTS.txt
  */
 
-import './elements/shack-app.js';
 import * as bench from '/bench.js';
-import {categories} from './data.js';
-
 bench.start();
-const app = document.createElement('shack-app');
-app.page = 'mens_tshirts';
-app.categories = categories;
-document.body.appendChild(app);
-requestAnimationFrame(() => bench.stop());
