+++ polymer/shack/index-2.js
@@ -1,0 +1,21 @@
+/**
+ * @license
+ * Copyright (c) 2019 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt The complete set of authors may be found
+ * at http://polymer.github.io/AUTHORS.txt The complete set of contributors may
+ * be found at http://polymer.github.io/CONTRIBUTORS.txt Code distributed by
+ * Google as part of the polymer project is also subject to an additional IP
+ * rights grant found at http://polymer.github.io/PATENTS.txt
+ */
+
+import './elements/shack-app.js';
+import * as bench from '/bench.js';
+import {categories} from './data.js';
+
+const app = document.createElement('shack-app');
+app.page = 'mens_tshirts';
+app.categories = categories;
+document.body.appendChild(app);
+
+requestAnimationFrame(() => bench.stop());
